requirements.txt
langchain
faiss-cpu
sentence-transformers
streamlit
pypdf
python-dotenv
tqdm